fix(net/rdr3): Deadlock in CProcessControlUpdateGraph
During the initial port of Onesync to RedM, A cover related update function responsible for updating players lastCover position and some grid information used within the cover system was accidently removed in onesync leading to a common deadlock when the player navigates from the center of the map (0.0, 0.0, 0.0) to surrounding coordinates/areas

This patch removes the hook restoring the intended functionality and resolving the Deadlock. The function that was intended to be nop (similar to how it was in FiveMs OneSync implementation) does not seem to be present in RDR3.